requirements.txt
algebraic-data-types==0.2.1
amqp==5.1.1
appnope==0.1.0
asgiref==3.5.2
asttokens==2.1.0
async-timeout==4.0.2
autopep8==1.4.4
aws-psycopg2==1.1.1
backcall==0.1.0
beautifulsoup4==4.8.0
billiard==3.6.4.0
blessed==1.15.0
blis==0.7.9
#botocore==1.12.238
cached-property==1.5.1
catalogue==2.0.8
celery==5.2.7
cement==2.8.2
certifi==2019.3.9
cffi==1.15.1
chardet==3.0.4
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.3.9
confection==0.0.3
contourpy==1.0.6
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.4
cron-descriptor==1.2.32
cronsim==2.3
cryptography==38.0.3
cycler==0.11.0
cymem==2.0.7
Cython==0.29.32
decorator==4.4.0
defusedxml==0.6.0
Deprecated==1.2.13
dj-database-url==1.0.0
dj-email-url==1.0.6
Django==4.1.3
django-allauth==0.39.1
django-appconf==1.0.5
django-bulk-update==2.2.0
django-cache-url==3.4.2
django-compressor==4.1
django-constance==2.9.1
django-cors-headers==2.1.0
django-countries==5.3.3
django-crispy-forms==1.7.2
django-cryptography==1.1
django-debug-toolbar==1.10.1
django-extensions==3.1.5
django-filter==2.4.0
django-htmlmin==0.11.0
django-js-asset==1.2.2
django-mptt==0.10.0
django-picklefield==1.0.0
django-ratelimit==2.0.0
django-rq==2.6.0
django-taggit==3.1.0
django-urlconf-export==1.1.1
django-widget-tweaks==1.4.12
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==3.2.3
dlib==19.24.0
#docker==3.7.3
#docker-compose==1.23.2
#docker-pycreds==0.4.0
#dockerpty==0.4.1
docopt==0.6.2
docutils==0.15.2
drf-extensions==0.3.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.4.1/en_core_web_sm-3.4.1-py3-none-any.whl
executing==1.2.0
ExifRead==3.0.0
face-recognition==1.3.0
face-recognition-models==0.3.0
fido2==1.1.0
fonttools==4.38.0
frozendict==2.3.5
future==0.16.0
geographiclib==2.0
geopy==2.3.0
graphql-core==3.2.3
graphql-relay==3.2.0
html5lib==1.0.1
idna==2.8
#ipdb==0.13.9
#ipython==7.8.0
#ipython-genutils==0.2.0
itypes==1.2.0
jaeger-client==4.8.0
jedi==0.15.1
Jinja2==3.1.2
jmespath==0.9.4
joblib==1.2.0
jsonrpclib-pelix==0.4.3.2
jsonschema==2.6.0
kiwisolver==1.4.4
kombu==5.2.4
langcodes==3.3.0
Markdown==3.1.1
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
mistune==0.8.4
murmurhash==1.0.9
networkx==2.8.8
nltk==3.7
numpy==1.23.5
oauthlib==3.0.1
opencv-python==4.6.0.66
opentracing==2.4.0
packaging==21.3
pandas==1.5.2
parso==0.5.1
pathspec==0.5.9
pathy==0.8.1
pep8==1.7.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==9.3.0
preshed==3.0.8
prompt-toolkit==2.0.9
psycopg2==2.9.5
ptyprocess==0.6.0
pure-eval==0.2.2
pycocotools==2.0.6
pycodestyle==2.5.0
pycparser==2.21
pycurl==7.45.1
pydantic==1.10.2
pygame==2.3.0
Pygments==2.4.2
PyJWT==1.5.3
pyocclient==0.6
pyotp==2.7.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-decouple==3.1
python3-openid==3.1.0
pytz==2021.3
PyYAML==3.13
rcssmin==1.1.0
redis==4.3.4
regex==2022.10.31
requests==2.21.0
requests-oauthlib==1.2.0
rjsmin==1.2.0
rq==1.11.1
scikit-learn==1.1.3
scipy==1.9.3
seaborn==0.12.1
segno==1.5.2
semantic-version==2.5.0
#sentry-sdk==1.11.0
serpy==0.3.1
six==1.11.0
smart-open==5.2.1
soupsieve==1.9.3
spacy==3.4.3
spacy-legacy==3.0.10
spacy-loggers==1.0.3
sqlparse==0.2.4
srsly==2.4.5
stack-data==0.6.1
statsd==4.0.1
stripe==2.27.0
tabulate==0.9.0
termcolor==1.1.0
texttable==0.9.1
thinc==8.1.5
threadloop==1.0.2
threadpoolctl==3.1.0
thrift==0.16.0
toml==0.10.2
torch==1.13.0
torchvision==0.14.0
tornado==6.2
tqdm==4.64.1
traitlets==4.3.2
typer==0.7.0
typing_extensions==4.4.0
uritemplate==4.1.1
urllib3==1.24.2
vine==5.0.0
wasabi==0.10.1
wcwidth==0.1.7
webencodings==0.5.1
websocket-client==0.56.0
whitenoise==6.2.0
wrapt==1.14.1
z3-solver==4.12.1.0
